requirements.txt
annoy==1.17.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
emoji==1.2.0
fastapi==0.65.1
firebase-admin==5.0.0
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googlemaps==4.4.5
grpcio==1.37.1
gunicorn==20.1.0
h11==0.12.0
httplib2==0.19.1
httptools==0.1.2
idna==2.10
Janome==0.4.1
joblib==1.0.1
mojimoji==0.0.11
msgpack==1.0.2
neologdn==0.4
numpy==1.20.3
packaging==20.9
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
smart-open==5.0.0
starlette==0.14.2
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
